Continuação do staff e inicio do back-office-user.

+++ src/app/components/back-office-agent/listing-list/listing-list.component.ts
@@ -57,17 +57,17 @@ export class ListingListComponent {
   }
 
   deleteList(id: number) {
-
-    this.listingService.deleteListing(id).subscribe({
-      next: () => {
-        window.location.reload();
-        console.log('Listing deleted successfully');
-      },
-      error: (err) => {
-        console.error('Error deleting listing:', err);
-      }
-    });
-
+    if (confirm("Tem a certeza que quer apagar este imóvel?")) {
+      this.listingService.deleteListing(id).subscribe({
+        next: () => {
+          window.location.reload();
+          console.log('Listing deleted successfully');
+        },
+        error: (err) => {
+          console.error('Error deleting listing:', err);
+        }
+      });
+    }
   }
